website: attempt to dissuade use of consul_service
The consul_service resource is not something most users should use, since service registration is normally done by a Consul Agent once a machine is running rather than via an external API call during provisioning.

The name of this resource makes it a bit of an attractive nuisance for those intending instead to register "External Services" as described in the Consul documentation. For that, the consul_catalog_entry resource is the right approach since it allows creation of a service entry that isn't associated with a "real" node.