consul_catalog_service: allow "tag" and "datacenter" to be set
Previously these attributes were incorrectly flagged as `Computed`, probably due to an oversight when adapting the schema of the corresponding resource.

They need to be marked as `Optional` so that they can actually be set in configuration, which we now do here. The expected read behavior was already present in spite of the incorrect schema, as verified by the new test.

This fixes #29.