resource/consul_service: refactor to use catalog
This moves the consul_service resource to the catalog APIs for service
registration and deregistration. Agent APIs are intended
for service management on-node, which Terraform is not
designed for.

This requires 'node' be set for each service, and for the node to exist.
It will not allow an upsert, as is the behavior of the Consul API.