Default datacenter to "" if not readable in /agent/self
Most resources uses the datacenter of the agent Terraform is connected
as default. This cause issues when Terraform does not have the agent:read
authorization.

For most operations, we can use an empty string and Consul will use the
datacenter of the agent being queried.

Fix https://github.com/terraform-providers/terraform-provider-consul/issues/97