Fix error checking in resourceConsulACLPolicyRead()
Failing to read a policy from the server does not necessarly mean
that the policy has been removed, the network can be down, the correct
datacenter may not be reachable etc. We must be conservative when
removing resources from the state and only create a new one if it's
actualy needed.

See https://github.com/terraform-providers/terraform-provider-consul/issues/160