Use a temporary workaround to set the Content-Type header
This define a custom RoundTripper to add the Content-Type header when
it is needed. This commit should be reverted once https://github.com/hashicorp/consul/pull/10204
is merged and released upstream.

I did not write a test for this change but tested locally and the
header is correctly present for PUT requests but not for GET or DELETE.