Add support for exported-services and mesh config entries
This path add support for the two new config entriy kinds.

I've encountered an issue that is making it impossible to run the test
against a running Consul server as we were doing until now: after
creating an Admin partition it is not possible to create network areas,
which means that the tests can fail based on the order in which they ran.

Since we cannot control that order in Go, I've had to change strategy
and start a new Consul server for each test. That way all tests are
completely isolated and there is no more risks of collision between
tests.

There is however the risk that the server is not completely initialised
before running the test. I've put some `time.Sleep()` calls in
`startServerWithConfig()` to work around this for now but it is an
ugly solution and I will have to improve this later.