Fix handling of partition in consul_config_entry
Having the partition argument in `config_json` was causing an issue as
the provider was not reading the result back from the correct partition.
This is now fixed by having `partition` as an actual argument of
`consul_config_entry`.

Also, fix some flaky tests by improving the tests in startTestServer().

Also, simplify local testing by adding an install target in the Makefile.