Fix handling of partition in consul_config_entry (#300)
Having the partition argument in `config_json` was causing an issue as
the provider was not reading the result back from the correct partition.
This is now fixed by having `partition` as an actual argument of
`consul_config_entry`.

Also, fix some flaky tests by improving the tests in startTestServer().

Also, simplify local testing by adding an install target in the Makefile.

Also, skip remote datacenters in Docker as they are failing. I think there is some networking issue and will look into it later.