Add read support
This patch adds a new ConfigEntryImplementation interface that can be
used to implement all the concrete config entries.

The ConfigEntryImplementation needs to implement a `Decode()` method to
decode the config entry from the user configuration and a `Write()` to
update the state during refresh. What's needed in addition to those two
methods will be taken care of by `resourceFromConfigEntryImplementation()`.