Implement the consul_config_entry_sameness_group resource
This resource makes it easier to manage a `sameness-group` config entry.
It builds on https://github.com/hashicorp/terraform-provider-consul/pull/364
so should be reviewed and merged after it.

The importer implementation in `resourceFromConfigEntryImplementation()`
has also been updated to support the `"<name>" or "<partition>/<name>"`
ID format has `sameness-group` is not bound to a namespace.